Features completed: Added geofencing algorithm and alarm system added vibration Changed the tracker to only set off the alarm after all data points have been collected for the geofencing
Errors Fixed:
Alarm goes off with a too small of a radius. (Server side needs to fix)
Partially fixed error when the user turns off the volume. (Now it sets the ringer volume back up)

Current Errors:
Work on sending emails when user forgets their password. (added but attempting to still send email without opening gmail)
Finish the turn on/off feature (add a feature to not be able to turn off the tracking when the alarm is going off)
When user clicks the volume button during alarm sound, user can still turn down the volume

To complete:
After user enters new contacts, save the contacts to a data base
When registering new accounts, update the data base.
Save the users points to a text file.
Work on getting the camera to take pictures or record video.
Have a feature to look at the pictures/videos that were take.
Have a feature to look at the media that was recorded from the sound.
Work on sending text messages